fix: EvenementRepository + GroupeRepository Integer -> Long
Les models Evenement et Groupe utilisent Long pour l'ID (@GeneratedValue Long id)
mais les repositories etaient declares avec Integer comme type de cle.
Correction : JpaRepository<Evenement, Long> et JpaRepository<Groupe, Long>

+++ src/main/java/com/dembaandousmane/gest_priere/evenement/repository/EvenementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EvenementRepository extends JpaRepository<Evenement, Integer> {
+public interface EvenementRepository extends JpaRepository<Evenement, Long> {
 
 }
